MapActivity: Open QRDetailsActivity when QR on map is clicked
Since the QR code score is shown in the marker info window right now,
the event listener is implemented on that. However, ideally the marker
itself would display the score without having to click on it first to
display the info window, after which the handler code can be moved to
the correct listener.